requirements/base.txt
configparser==3.5.0
Django==1.11
django-allauth==0.25.2
django-crispy-forms==1.6.0
django-countries==4.0
django-dotenv==1.4.1
django-filter==0.15.3
djangorestframework==3.5.2
djangorestframework-jwt==1.8.0
enum34==1.1.6
fake-factory==0.5.7
flake8==3.0.4
inflection==0.3.1
ipaddress==1.0.17
Markdown==2.6.6
mccabe==0.5.2
mixer==5.5.7
pep8-naming==0.4.1
Pillow==3.3.1
psycopg2==2.6.2
pycodestyle==2.0.0
pydocstyle==1.1.1
pyflakes==1.2.3
PyJWT==1.4.2
python-dateutil==2.5.3
qrcode==5.3
six==1.10.0
django-cors-headers==1.3.0
redis==2.10.5
django-q==0.7.18
django-paypal==0.3.3
django-colorfield==0.1.14
